Reorganize repo
CocoaPods needs to be able to refer to specific groups of files for
different platforms. The easiest way to make this happen is to move iOS
and Mac source files into their own directories, and then to have a
shared directory with source files that are used on both platforms.

I'm also moving the Frank specific files into their own directory to
separate them from the rest of the more generic source files.